+++ CHANGELOG.md
@@ -16,3 +16,9 @@
 - New workflow to confirm builds are passing on supported Node.js versions.
 - Dependency updates.
 - Add LIMIT_REDUCE_ONLY order type.
+
+## 1.2.0
+
+- Now exposing the public client that is created when the SDK is initialised.
+- Now exporting typings in their own import @ `rysk-sdk/types`.
+- Constants are now exposed via the client. This includes the API URL, Chain and available margin assets.